search now works and returns results.
modified getMoviesByQuery to pass the serach query and page number.

changed image size in upcomingMovies to 154 and on Carousel to 342 to make the site load faster.

header and footer are now fixed .

added api key to .env file.

additional css changes to fix small rendering issues.

code is stable and works, no bugs or error messages when i use it.